app/publisher/requirements.txt
Flask==3.0.3
google-cloud-pubsub==2.20.2
google-cloud-secret-manager==2.20.2
gunicorn==22.0.0
